Created AuthState
