+++ app/App.tsx
@@ -2,20 +2,20 @@ import 'react-native-dotenv';
 import 'react-native-config';
 import React from 'react';
 // import {SafeAreaView, ScrollView, StatusBar} from 'react-native';
-import GlobalState from './src/Context/GlobalState';
 import {NavigationContainer} from '@react-navigation/native';
 import MainNavigator from './src/Navigation/MainNavigator';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
+import AuthState from './src/Context/AuthState';
 
 const RightPayApp = () => {
   return (
-    <GlobalState>
+    <AuthState>
       <SafeAreaProvider>
         <NavigationContainer>
           <MainNavigator />
         </NavigationContainer>
       </SafeAreaProvider>
-    </GlobalState>
+    </AuthState>
   );
 };
