+++ app/src/types/AppContextType.ts
@@ -1,12 +1,11 @@
 import {CreditCard} from './CreditCardType';
-import { Location } from './Location';
-import {Profile} from './ProfileType';
+import {Location} from './Location';
 
 export interface AppContext {
   creditCards: CreditCard[];
   addNewCreditCard: (_creditCard: CreditCard) => void;
   removeCreditCard: (_creditCard: CreditCard) => void;
-  userProfile: Profile;
-  setUserProfile: (_profile: Profile) => void;
-  location:Location;
+  location: Location;
+  isLoading: boolean;
+  setIsLoading: (_isLoading: boolean) => void;
 }
